correção de bugs funcionalidade 1
